Modularize the app. start fix dependencies.
