+++ main.py
@@ -1,0 +1,10 @@
+# main.py
+
+from app import app
+from routes import *
+
+def main():
+    app.run()
+
+if __name__ == "__main__":
+    main()
