+++ util/start_server_persist
@@ -13,7 +13,7 @@ while [[ 1 -eq 1 ]]
 do
 echo "STARTING Flask Server"
 #echo "  Press CTRL+C TWICE to QUIT"
-python app.py
+python main.py
 
 restcount=10
 echo -e "\nPERSIST SERVER:"
